deleted ariel-test (refactoring changes)
