--- ariel-test/.wpilib/wpilib_preferences.json
@@ -1,6 +1,0 @@
-{
-    "enableCppIntellisense": false,
-    "currentLanguage": "java",
-    "projectYear": "2022",
-    "teamNumber": 3566
-}
