--- ariel-test/src/main/java/frc/robot/commands/ExtendIntake.java
@@ -1,36 +1,0 @@
-package frc.robot.commands;
-
-import frc.robot.subsystems.IntakeSubsystem;
-import edu.wpi.first.wpilibj2.command.CommandBase;
-
-public class ExtendIntake extends CommandBase {
-  @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-
-  /**
-   * Creates new extend intake command.
-   *
-   * @param subsystem The subsystem used by this command.
-   */
-  public ExtendIntake(IntakeSubsystem subsystem) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements();
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {}
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {}
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {}
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
