+++ src/main/java/frc/robot/RobotContainer.java
@@ -10,18 +10,28 @@ import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 
 import frc.robot.subsystems.DriveSubsystem;
+import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.commands.DriveWithJoystick;
+import frc.robot.commands.IntakeCommand;
 
 public class RobotContainer {
-  private Joystick JS = new Joystick(0);
 
-  private final DriveSubsystem driveSubsystem = new DriveSubsystem("SparkMax");
+  // Inputs
+  public Joystick js1 = new Joystick(0);
   
-  private final DriveWithJoystick driveWithJoystick = new DriveWithJoystick(JS, driveSubsystem);
+  // Subsystems
+  private final DriveSubsystem driveSubsystem = new DriveSubsystem("SparkMax");
+  private final IntakeSubsystem intakeSubsystem = new IntakeSubsystem();
 
+  // Commands
+  private final DriveWithJoystick driveWithJoystick = new DriveWithJoystick(js1, driveSubsystem);
+  private final IntakeCommand intakeCommand = new IntakeCommand(intakeSubsystem);
+
+  /** The container for the robot. Contains subsystems, OI devices, and commands. */
   public RobotContainer() {
-    driveSubsystem.setDefaultCommand(driveWithJoystick);
+    // Configure the button bindings
     configureButtonBindings();
+    driveSubsystem.setDefaultCommand(driveWithJoystick);
   }
 
   /**
@@ -30,7 +40,11 @@ public class RobotContainer {
    * edu.wpi.first.wpilibj.Joystick} or {@link XboxController}), and then passing it to a {@link
    * edu.wpi.first.wpilibj2.command.button.JoystickButton}.
    */
-  private void configureButtonBindings() {}
+  private void configureButtonBindings() {
+    // Intakes object (in)
+    JoystickButton js1_b12 = new JoystickButton(js1, 12);
+    js1_b12.whenHeld(intakeCommand);
+  }
 
   /**
    * Use this to pass the autonomous command to the main {@link Robot} class.
