+++ src/main/java/frc/robot/commands/IntakeCommand.java
@@ -47,5 +47,5 @@ public class IntakeCommand extends CommandBase {
   public boolean isFinished() {
     return false;
   }
-  
+
 }
